@@ -1,32 +1,46 @@
-window.$ = window.jQuery = require('jquery');
-require('bootstrap-sass');
-
+import $ from 'jquery';
+import _ from 'lodash';
+import 'bootstrap-sass';
 import Vue from 'vue';
-import VueResource from 'vue-resource';
+import axios from 'axios';
+
+window.$ = window.jQuery = $;
+window.axios = axios;
+
 import router from './routes';
 import store from './store/index';
 import App from './components/App.vue';
 import jwtToken from './helpers/jwt-token';
 
-Vue.use(VueResource);
-Vue.component('app', App);
-
-Vue.http.interceptors.push((request, next) => {
-    request.headers.set('X-CSRF-TOKEN', Laravel.csrfToken);
+axios.interceptors.request.use(config => {
+    config.headers['X-CSRF-TOKEN'] = window.Laravel.csrfToken;
+    config.headers['X-Requested-With'] = 'XMLHttpRequest';
 
     if(jwtToken.getToken()) {
-        request.headers.set('Authorization', 'Bearer '+ jwtToken.getToken());
+        config.headers['Authorization'] = 'Bearer '+ jwtToken.getToken();
     }
 
-    next((response) => {
-        if(!response.ok && response.body.error && (response.body.error === "token_invalid" || response.body.error === "token_expired" || response.body.error === 'token_not_provided')) {
-            store.dispatch('logoutRequest')
-                .then(()=> {
-                    router.push({name: 'login'});
-                });
-        }
-    });
+    return config;
+}, error => {
+    return Promise.reject(error);
 });
+
+axios.interceptors.response.use(response => {
+    return response;
+}, error => {
+    let errorResponseData = error.response.data;
+
+    if(errorResponseData.error && (errorResponseData.error === "token_invalid" || errorResponseData.error === "token_expired" || errorResponseData.error === 'token_not_provided')) {
+        store.dispatch('logoutRequest')
+            .then(()=> {
+                router.push({name: 'login'});
+            });
+    }
+
+    return Promise.reject(error);
+});
+
+Vue.component('app', App);
 
 const app = new Vue({
     router,
