@@ -1,16 +1,17 @@
 <template>
     <div class="container">
         <div class="jumbotron">
-            <h1>Welcome to <span v-once>{{siteName}}</span></h1>
-            <p>Starter Boilerplate SPA made with Laravel 5.3, JWT Auth, Vue 2, Vue-router 2, Vuex 2 and love</p>
+            <h1 v-once>{{siteName}}</h1>
+            <p>Starter Boilerplate SPA made with Laravel 5.4, JWT Auth, Vue 2, Vue-router 2, Vuex 2, Axios and love</p>
         </div>
     </div>
 </template>
 
 <script>
-    import {siteName} from './../config';
+    import {siteName} from './../../config';
 
-    export default{
+    export default {
+        name: 'home',
         data() {
             return {
                 siteName: siteName
