@@ -14,6 +14,6 @@ return [
     */
 
     'previous' => '&laquo; Previous',
-    'next'     => 'Next &raquo;',
+    'next' => 'Next &raquo;',
 
 ];
